Changed free_rr_data() so it sets the rr_data_t pointer to NULL.
Since i got repeated segmentation faults because free_rr_data() tried
to free "random" memory it could be that an already freed pointer
was given to the function. Maybe this change reveals a bit more of the
real point where the error is because it runs into NULL pointer
dereference earlier.