Fix checking of array against NULL
Detected by Coverity (CID 291937). Checking if an array is not NULL is always true. It seems that the arrays were pointers some time ago and it was forgotten to fix these checks. Since these arrays contain binary data we can only be sure that they are empty/not set when comparing the complete array against zero. This is done now.