Use more secure random numbers.
Add function getrandom64() which first tries to get good/safe random numbers and then falls back to less good/safe random numbers if it is not possible.
Since it must work for different operating systems the function maybe is not the most elegant one but it should work for now.