build(configure): Better test if compiler is suitable (#66)
Add a simple check if the compiler finds an important header file. If not it is not really usable. This fixes an issue with Cygwin that could find Clang that is installed on Windows but not for/on Cygwin (by default this is the case on the AppVeyor machines) and thus does not include all the Linux stuff that  is necessary.