Refactor systemd user unit to be a proper user unit
This commit refactors the user scoped systemd unit to be a proper
user scoped unit as opposed to the common "hack" of a system scoped
unit with the user taken as a parameter.

Additionally, this adjusts the behavior of both systemd units to
pass parameters specified in the OPTARGS parameter in /etc/sysconfig/cntlmd
to both user as well as system instances for purposes of centralized management
of parameters such as the proxy hostnames or PAC file while still allowing
users to maintain their credentials in a configuration file private to their homedir.