Refactor of packaging files + migrate from sysvinit to systemd service management
* Relocate all packaging files under linux/<distro>
* Use new specfile derived from the Fedora EPEL upstream
* Build with debugging symbols at all times, such that we can strip them into the associated devel/debuginfo package during packaging.
* Updates to README as needed
* Use native systemd units for service management

This commit implements both a user scoped systemd unit in order to support use cases where each user maintains one's own configuration file, running cNTLM as one's own user, as well as the traditional system scoped unit that is prevalent in most distribution packages. This is more compatible (as it does not require user scoped systemd instances) however less flexible.

Additionally, the behavior of both systemd units is designed to pass parameters specified in the OPTARGS parameter in /etc/sysconfig/cntlmd to both user as well as system instances of cntlm, for purposes of centralized management of configuration parameters, such as the proxy hostnames or PAC file, while still allowing users to maintain their credentials in a configuration file private to their homedir.